serialize errors and fix a long-standing hopefully
There was a unique edge case where Remastered useRouteData could throw unknown errors.
This happened when:
* User visited `/page-a`
* User navigated to `/page-b`
* User clicked back, and returned to `/page-a`
* User refreshed
* _Before complete hydration_, user clicks forward, visiting page `/page-b`

What this did was made Remastered anticipate for values for `/page-b` in the initial server
payload, while it had `/page-a`, so we had no values, hence the error.

This is now fixed by forcing the location state from the server, AKA `/page-a`,
and after hydration we listen to history changes and apply them accordingly.